make lessons topbar sticky
