+++ src/js_modules/syllabus/ProjectBoardGuidedExperience.jsx
@@ -3,6 +3,9 @@ import React, { useState, useRef, useEffect } from 'react';
 import useTranslation from 'next-translate/useTranslation';
 import PropTypes from 'prop-types';
 import { Box, Button, useColorModeValue } from '@chakra-ui/react';
+import SubtasksPill from './SubtasksPill';
+import StatusPill from './StatusPill';
+import Topbar from './Topbar';
 import TaskCodeRevisions from './TaskCodeRevisions';
 import OpenWithLearnpackCTA from './OpenWithLearnpackCTA';
 import useModuleHandler from '../../common/hooks/useModuleHandler';
@@ -11,70 +14,6 @@ import ReactPlayerV2 from '../../common/components/ReactPlayerV2';
 import Heading from '../../common/components/Heading';
 import Text from '../../common/components/Text';
 import Icon from '../../common/components/Icon';
-
-function SubtasksPill() {
-  const { t } = useTranslation('common');
-  const { subTasks } = useModuleHandler();
-
-  if (!Array.isArray(subTasks) || subTasks.length === 0) return null;
-
-  const tasksDone = subTasks.length > 0 && subTasks?.filter((subtask) => subtask.status === 'DONE');
-
-  return (
-    <Box padding="5px 7px" fontSize="13px" borderRadius="27px" border="1px solid #0084FF" color="#0084FF">
-      {`${tasksDone.length} / ${subTasks.length} ${t('tasks')}`}
-    </Box>
-  );
-}
-
-function StatusPill() {
-  const { t } = useTranslation('syllabus');
-  const { currentTask } = useModuleHandler();
-  const { hexColor } = useStyle();
-
-  if (!currentTask || currentTask.task_status === 'PENDING') return null;
-
-  const colorsDict = {
-    APPROVED: {
-      background: hexColor.greenLight2,
-      color: hexColor.greenLight,
-    },
-    REJECTED: {
-      background: 'red.light',
-      color: hexColor.danger,
-    },
-    PENDING: {
-      background: 'yellow.light',
-      color: hexColor.yellowDefault,
-    },
-    IGNORED: {
-      background: hexColor.greenLight,
-      color: hexColor.green,
-    },
-  };
-
-  const labelsDict = {
-    APPROVED: t('approved'),
-    PENDING: t('pending'),
-    IGNORED: t('approved'),
-    REJECTED: t('rejected'),
-  };
-
-  const revisionStatus = currentTask.revision_status;
-
-  return (
-    <Box
-      padding="5px 7px"
-      borderRadius="27px"
-      background={colorsDict[revisionStatus]?.background}
-      color={colorsDict[revisionStatus]?.color}
-      fontWeight="500"
-      fontSize="13px"
-    >
-      {labelsDict[revisionStatus]}
-    </Box>
-  );
-}
 
 function ProjectHeading({ currentAsset, isDelivered }) {
   const { backgroundColor4, hexColor } = useStyle();
@@ -239,34 +178,7 @@ function ProjectBoardGuidedExperience({ currentAsset }) {
           <TaskCodeRevisions />
         )}
       </Box>
-      <Box
-        id="project-topbar"
-        zIndex="20"
-        background={backgroundColor4}
-        position="sticky"
-        top="-1px"
-        // margin={{ base: '-1.5rem -10px 0 -10px', md: '-1.5rem -2rem 0 -2rem' }}
-        borderBottom="1px solid #BBE5FE"
-        padding="15px"
-        display={isHeaderVisible ? 'none' : 'flex'}
-        justifyContent="space-between"
-        alignItems="center"
-      >
-        <Box display="flex" alignItems="center">
-          <Icon icon={assetTypeIcons[assetType] || 'book'} width="20px" height="20px" color={hexColor.blueDefault} style={{ margin: 'auto', marginRight: '0.4rem' }} />
-          <Heading style={{ fontWeight: '400' }} size="xsm" display="inline-flex" gridGap="10px" margin="0 0 0 0 !important">
-            {title}
-          </Heading>
-        </Box>
-        <Box display="flex" alignItems="center" gap="5px">
-          <StatusPill />
-          <SubtasksPill />
-          <Button display="flex" alignItems="center" gap="5px" variant="ghost" color={hexColor.blueDefault} onClick={scrollTop}>
-            <Icon icon="arrowLeft2" style={{ transform: 'rotate(90deg)' }} color={hexColor.blueDefault} />
-            {t('back-to-top')}
-          </Button>
-        </Box>
-      </Box>
+      <Topbar currentAsset={currentAsset} display={isHeaderVisible ? 'none' : 'flex'} />
     </>
   );
 }
@@ -288,3 +200,5 @@ ProjectHeading.defaultProps = {
 };
 
 export default ProjectBoardGuidedExperience;
+
+export { StatusPill, SubtasksPill };
